moved gizmo handler to separate file
